requirements.txt
Django>=4.2
djangorestframework
django-cors-headers
celery>=5.3
redis>=5
django-celery-beat>=2.6
croniter>=2.0